Add search submit and clear buttons to search field
This feature will wrap the search field in an input group and
append one or both of the following buttons: submit and clear.
These buttons are self explanatory: searchButton will do exactly what
'searchOnEnterKey' does and clearSearchButton will clear the
existing search field text (and may issue a data reload).

This may be possible with all of the new themes. This commit
handles bootstrap 3 and 4 plus bulma (but only bootstrap 4 has been
tested).